+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.27.3...HEAD) - TBD
 
+### :bug: Fixed
+
+- Use `example`/`examples` declared on array items in the examples phase.
+
 ## [4.27.3](https://github.com/schemathesis/schemathesis/compare/v4.27.2...v4.27.3) - 2026-09-17
 
 ### :rocket: Added